Fix sync of cleaning deployment state hash
Previously, adding things to the deployment
tracking hash was synchronized, but removing things
from it was not.

Since both methods are managing our global state, I added
a mutex to the deletion method.

It also adds asm/service_deployment as a require (b/c
asm should always be an entry point for that class b/c
it manages syncronizations so that service deployments
so not have to be thread safe)

The require could not be added to top scope b/c of
weird ruby namespace issues.