Add status check
The following patch adds a new rest service
status that can be used to retrieve the current status.

This status is determiend by parsing the last line that
starts with Status: XXX out of the deployment logs.

This patch also adds logs that indicate status at a high level.

It also moves the logging functionality to a Util method.o