skip server deployment if already complete
Currently we cannot rely on the idempotent nature of puppet apply for
the server swim lane because the idrac module always reboots the
server. For now we check to see if the server already has a razor
policy with the correct O/S attached to it and skip the server
deployment step if so.