Change inventory update to use controller rest call
Previously, we were just rerunning the puppet device
call for storage instances with no resources in order
to ensure things were re-inventoried in Puppet's database.

I was under the impression that this would work (and
was told that the controller would update itself based
on what inventory data was in Puppet or that it would
use that or something) Apparently, this is not the case,
and the inventory has to be done through this exact call.

I could not get this going b/c I have no idea what the
ref id is supposed to be. This patch assumes it is
the asmGUID, adds that as a parameter and uses that,
but it does not work and gets a 404. I also tried with
the puppet cert name, which also did not work.

The fact that our identifiers between the front end and
backend has been a constant source of pain for me. They
should be unified to save us a lot of effort and confusion.

I started to think about another issue related to this patch.
If there is an external web service that collects facts from
devices by running puppet device, this means that it could
be initiated at the same time as our puppet device calls which
can lead to issue. These calls need to be made in a thread safe
way, which means they have to move off the contoller.