Configure vswitch0 and user-specified hypervisor network
vSwitch0 and the user-specified hypervisor_network were not being
configured because they were conflicting with the default DHCP
Managment Network which is present on vSwitch0 after ESXi installation
via razor.

This patch configures vSwitch0 with vmnic0 and vmnic1 and adds the
user-specified hypervisor_management network as a second VMkernel port
group in addition to the default Management Network port group.

In the future we need to also remove the original Management Network,
remove the host from the cluster and add it back with the
user-specified hypervisor_netowrk static IP address.