Update how compellent retrieves wwns from the server
This commit refactors the code related to retrieving
server ids used by compellent.

1. It renames GetWWPN to ASM::WWPN

the namespace of that class did not even match its location
on disk, it was also called GetWWPN (a verb) instead of
WWPN (a noun/object).

The class contains lots of code for encrytion/decrytion
that should not be in the class (this code was removed)

The class was also simplified to have a single method called
get. This method was simplied to be just 2 lines of code.

2. It removes compellent specific logic that was contained
in process generic.

Process generic is only intended for common code, component
specific code should be moved to the component specific
processing methods.

It was also updating the file instead of updating the hash
before it gets written into the file. This made the code
way overly complicated and brittle. Changing it to just
adjust the resource hash, cut the amount of code required by
at least 5X.

3. It removed some copy paste code that did not actually seem
to do anything.

4. It adds a few comments to parts of the code that I found
to be unclear.