ASM-1307 fix excessive logging to asm_puppet.log
Debug messages were not being matched properly in ASM::Util.get_logs
resulting in a warning message in the asm_puppet.log file for every
debug message in the deployment logs. Because the ASM GUI polls the
logs this results in huge asm_puppet.log files. SST failed because the
file reached the maximum filesystem size of 32GB.

This commit matches the debug messages and adds the severity to the
returned JSON. Debug messages are filtered out because they are not
generally appropriate for display in the GUI.