Rewrote hyperv specific test
Previously, a hyperv test existed in the service deployer
that was failing and testing code that had been moved to
asm::processor::server.

This commit completely refactors the code so that it actually
does test the way that the hyper-v specific server code
operates.

It also adds a method to mock setting the components_by_type.