Fix logic to skip deployment with razor 0.13.0 api
process_server has logic to skip adding a razor policy to install an
O/S on a server if the server already has a matching policy. The razor
0.13.0 upgrade renamed installers to tasks. The razor upgrade patch
mistakenly looked for a policy parameter called "tasks" but it should
have been "task" (singular). This commit fixes that and updates the
corresponding spec test.