monitor razor O/S installation status
Monitor the progress of razor O/S installations by looking for related
events in the razor node log. This solves two problems:

1. If the razor O/S install fails for some reason, the node may boot
into an old O/S on its hard drive. This can cause very confusing
errors where the install appears to have succeeded, e.g. if the old
O/S is the same as the one we were attempting to install.

2. We make esxcli command-line calls to the ESXi host to determine if
it is up and ready to be added to vcenter. However, the ESXi install
requires two reboots, so we may see it as up, but then the host
reboots again. By monitoring the razor install log we can ensure that
it has booted fro the second time before we begin making esxcli calls.