fix device.conf decryption failure
The device.conf decryption code in asm_encode.rb doesn't work on
jruby, so it is executed by calling MRI ruby as a script. On recent
builds (not all the time?) this has started failing because jruby gems
are getting pulled in due to the GEM_HOME environment variable which
is set by torquebox. This fix executes the command with "env -i" to
run the command from a clean environment.