update vm network_interfaces parameter
The parameter used to be a comma-separated list of network
names. VMware used those as portgroup names to configure network
interfaces for on the VM. Hyper-V expected the names to include the
VLAN in them (e.g. names like VLAN27) and used those for
configuration.. The parameter is now passed as a list of network
hashes and the various VMs pull the appropriate data out of thme.