only wait for razor boot_install event for VMs
Our VMs have hard drive before PXE boot in the default boot
order. Because of that they do not check back in with razor after they
reboot and have bits installed on the hard drive. Therefore we cannot
wait for those events and instead have to rely on waiting for the
puppet agent to check in to validate completion of the O/S install.