Strip out the asm::idrac key from resource_hash if it exists
Arguably, the more elegant thing would be to edit all the templates and
the template generation logic to not assume that all servers will have
an 'asm::idrac' stanza in their config, and then add a dynamically
generated stanza based on the discovered idrac config.

However, that would be a story in its own right to implement, so we will
do the cheap and pragmatic thing for not and just kill the 'asm::idrac'
stanza before we pass node resources to Puppet for deployment.

Fixes ASM-1588