update ASM::NetworkConfiguration for rack server data
Rack server nic data is passed in an 'interfaces' array and blade
server nic data is passed in a 'fabric' array. NetworkConfiguration
creates a new field called 'cards' that contains the same nic data for
ease of parsing. It also strips out all disabled, non-applicable and
FC cards, ports and partitions for ease of parsing. The original data
is still available in the 'fabrics' or 'interfaces' array.

Additionally rack nic slots are mapped to interfaces in the order they
appear. E.g. if a rack server has nics in slots 3, 5 and 7 and the
configuration data specifies two cards, slots 3 and 5 will be used.