retries were rebooting server every time
Wrong parameter was being passed for O/S version. Also updated it to
wait for teh :bind event which happens earlier that :reboot (so we
will time out quicker).

Also, the block until task complete method was waiting for a specific
status which doesn't work if you're waiting for an earlier status (say
:bind instead of the normal terminal status like :boot_local_2). This
commit compares the statuses based on the order they occur.