find nodes based on hw_info rather than facts
Razor itself uses the hw_info hash to determine uniqueness, not the
facts. Relying on facts is problematic because they have the potential
to be overwritten. If a server is booted into the microkernel it will
continue to report its facts back to the IP address of the appliance
that it installed the microkernl with its node name. If you install a
new appliance that may overwrite the facts of an entirely different
node that gets discovered with the same node name.