ensure deployment backup directory creation is atomic
Migrate / scale-up / retry executions back up the original deployment
directories into sub-directories. Previously there was no guarantee
that was done atomically so there was the possibility of corruption in
those files. For example a backup directory might be created without a
deployment.json file. Subsequent executions rely on being able to read
the deployment.json file out of the previous backup directory and will
fail if it's not present.